refactor: Product 관련 클래스 product 패키지로 이동
