+++ src/main/java/cart/dto/ProductDto.java
@@ -1,6 +1,6 @@
 package cart.dto;
 
-import cart.domain.Product;
+import cart.domain.product.Product;
 
 public class ProductDto {
